@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.1+2
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,7 +39,7 @@ dependencies:
   flutter_bloc: ^8.0.1
   flutter_svg: ^1.0.3
   http: ^0.13.4
-  qr_code_scanner: ^0.7.0
+  mobile_scanner: ^2.0.0
   flutter_native_splash: ^2.1.5
   test: ^1.19.5
   bloc_test: ^9.0.0
