@@ -1,25 +1,17 @@
 import 'dart:developer';
-import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:qr_code_scanner/qr_code_scanner.dart';
+import 'package:mobile_scanner/mobile_scanner.dart';
 
 typedef OnResultScanned = Function(Barcode barcode);
-typedef OnScannedFinish = Function(Barcode barcode);
 
 class CustomQRView extends StatefulWidget {
-  final OnResultScanned? callback;
-  final OnScannedFinish? onFinish;
-  final QrScannerOverlayShape? overlay;
-  final EdgeInsetsGeometry overlayMargin;
+  final OnResultScanned? onResult;
 
-  const CustomQRView(
-      {Key? key,
-      this.callback,
-      this.onFinish,
-      this.overlay,
-      this.overlayMargin = EdgeInsets.zero})
-      : super(key: key);
+  const CustomQRView({
+    Key? key,
+    this.onResult,
+  }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => _QRViewState();
@@ -27,19 +19,9 @@ class CustomQRView extends StatefulWidget {
 
 class _QRViewState extends State<CustomQRView> {
   Barcode? result;
-  QRViewController? controller;
-  final GlobalKey qrKey = GlobalKey(debugLabel: 'QR');
-
-  // In order to get hot reload to work we need to pause the camera if the platform
-  // is android, or resume the camera if the platform is iOS.
-  @override
-  void reassemble() {
-    super.reassemble();
-    if (Platform.isAndroid) {
-      controller!.pauseCamera();
-    }
-    controller!.resumeCamera();
-  }
+  MobileScannerController? controller =
+      MobileScannerController(facing: CameraFacing.front, torchEnabled: true);
+  final GlobalKey qrKey = GlobalKey(debugLabel: 'QRCode');
 
   @override
   Widget build(BuildContext context) {
@@ -55,45 +37,23 @@ class _QRViewState extends State<CustomQRView> {
   Widget _buildQrView(BuildContext context) {
     // To ensure the Scanner view is properly sizes after rotation
     // we need to listen for Flutter SizeChanged notification and update controller
-    return QRView(
-      key: qrKey,
-      onQRViewCreated: _onQRViewCreated,
-      overlay: widget.overlay,
-      onPermissionSet: (ctrl, p) => _onPermissionSet(context, ctrl, p),
-    );
-  }
-
-  void _onQRViewCreated(QRViewController controller) {
-    setState(() {
-      this.controller = controller;
-    });
-    controller.scannedDataStream.listen((scanData) {
-      if (result == null) {
-        setState(() {
-          result = scanData;
+    // return QRView(
+    //   key: qrKey,
+    //   onQRViewCreated: _onQRViewCreated,
+    //   overlay: widget.overlay,
+    //   onPermissionSet: (ctrl, p) => _onPermissionSet(context, ctrl, p),
+    // );
+    return MobileScanner(
+        allowDuplicates: false,
+        controller: controller,
+        onDetect: (barcode, args) {
+          if (barcode.rawValue == null) {
+            debugPrint('Failed to scan Barcode');
+          } else {
+            debugPrint('Barcode found: ${barcode.toString()}');
+          }
+          widget.onResult?.call(barcode);
         });
-        widget.callback?.call(result!);
-        // Timer(1.seconds, () {
-        //   setState(() {
-        //     result = null;
-        //   });
-        // });
-        controller.pauseCamera();
-        controller.dispose();
-      }
-    }, onDone: () {
-      log("QR: onDone");
-      widget.onFinish?.call(result!);
-    });
-  }
-
-  void _onPermissionSet(BuildContext context, QRViewController ctrl, bool p) {
-    log('${DateTime.now().toIso8601String()}_onPermissionSet $p');
-    if (!p) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('no Permission')),
-      );
-    }
   }
 
   @override
