@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:moab_poc/design/colors.dart';
 import 'package:moab_poc/page/components/base_components/base_page_view.dart';
 import 'package:moab_poc/page/components/layouts/basic_header.dart';
 import 'package:moab_poc/page/components/layouts/basic_layout.dart';
 import 'package:moab_poc/page/components/qr_view.dart';
-import 'package:qr_code_scanner/qr_code_scanner.dart';
+import 'package:mobile_scanner/mobile_scanner.dart';
 
 import '../components/base_components/button/primary_button.dart';
 import '../components/base_components/button/simple_text_button.dart';
@@ -28,16 +27,17 @@ class AddChildScanQRCodeView extends StatelessWidget {
         content: Center(
           child: Stack(
             children: [
-              CustomQRView(
-                callback: (Barcode code) => _handleScanResult,
-                overlay: QrScannerOverlayShape(
-                    borderColor: MoabColor.scannerBoarder,
-                    borderRadius: 0,
-                    borderLength: 100,
-                    borderWidth: 2,
-                    cutOutSize: 200,
-                    overlayColor: Theme.of(context).backgroundColor,
-                    cutOutBottomOffset: 120),
+              Container(
+                width: double.infinity,
+                height: 300,
+                alignment: Alignment.center,
+                child: SizedBox(
+                  width: 200,
+                  height: 200,
+                  child: CustomQRView(
+                    onResult: (Barcode code) => _handleScanResult,
+                  ),
+                ),
               ),
               Padding(
                 padding: const EdgeInsets.fromLTRB(0, 200, 0, 0), // TODO
@@ -66,7 +66,7 @@ class AddChildScanQRCodeView extends StatelessWidget {
   }
 
   _handleScanResult(Barcode code) {
-    print('Scanned code: ${code.code ?? ''}');
+    print('Scanned code: ${code.rawValue ?? ''}');
     onNext();
   }
 }
